fix: rename helpers to check-utils, replace handle-error file
