+++ auth/src/check-utils/check-utils.auth-user.ts
@@ -1,6 +1,6 @@
 import { User } from '../auth.interfaces'
 
 export const checkAuthUser = (usersList: User[]): boolean => {
-  const userIsAuth: boolean = usersList.some((user) => user.isAuth === true)
-  return userIsAuth
+  const isAuthUser: boolean = usersList.some((user) => user.isAuth === true)
+  return isAuthUser
 }
