+++ auth/src/check-utils/check-utils.index-of-auth-user.ts
@@ -1,0 +1,7 @@
+import { User } from '../auth.interfaces'
+
+export const checkIndexOfAuthUser = (usersList: User[]): number => {
+  const indexOfAuthUser: number = usersList.findIndex((user) => user.isAuth === true)
+
+  return indexOfAuthUser
+}
