+++ auth/src/handle-error.ts
