--- auth/src/helpers/helpers.check-auth-user-index.ts
@@ -1,7 +1,0 @@
-import { User } from '../auth.interfaces'
-
-export const checkAuthUserIndex = (usersList: User[]): number => {
-  const authUserIndex: number = usersList.findIndex((user) => user.isAuth === true)
-
-  return authUserIndex
-}
