--- auth/src/helpers/index.ts
@@ -1,3 +1,0 @@
-export * from './helpers.handle-error'
-export * from './helpers.check-auth-user'
-export * from './helpers.check-auth-user-index'
